# Warmup diagnostics before ensure in sierra artifact compilation